requirements.txt
certifi==2019.9.11
Click==7.0
numpy==1.17.3
opencv-contrib-python==4.1.1.26
opencv-python==4.1.1.26
Pillow==6.2.1
wincertstore==0.2
